[Patrick] - #292: moves published_at field  out of metadata
This fixes a bug in which published_at field was being converted to
a String every time a news was updated.
This was happening because on news updates all content inside metadata
object is saved at once. Because published_at is a complex object we
would need to recreate it from its String representation that comes in
the http request before updating the news' metadata field in mongo.
To avoid this extra complexity and since published_at is never
handled in frontend. It is better to move it out of metadata completely.