[Patrick] - #153: configures TCP keepAlive to 2 minutes to avoid connection errors
http://mongoosejs.com/docs/connections.html says that long running
applications should consider configuring keepAlive.

https://docs.mongodb.org/manual/faq/diagnostics/#does-tcp-keepalive-time-affect-mongodb-deployments
says that "For MongoDB, you will have better results with shorter [than
OS's default] keepalive periods, on the order of 120 seconds (two minutes)."

http://tldp.org/HOWTO/TCP-Keepalive-HOWTO/overview.html says
"Keepalive is non-invasive, and in most cases, if you're in doubt, you
can turn it on without the risk of doing something wrong. But do
remember that it generates extra network traffic, which can have an
impact on routers and firewalls."

For our case it seems that we can benefit without risks, so lets use 2
minutes.